# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Code Review Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pull-requests:
    get:
      operationId: listPullRequests
      tags: [pullRequests]
      summary: List all pull requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, merged]
        - name: repositoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of pull requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}:
    get:
      operationId: getPullRequest
      tags: [pullRequests]
      summary: Get a pull request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pull request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}/trigger-review:
    post:
      operationId: triggerReview
      tags: [pullRequests]
      summary: Trigger AI review for a pull request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Review triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewJob"
  /pull-requests/jobs/{jobId}:
    get:
      operationId: getReviewJob
      tags: [pullRequests]
      summary: Get progress of a running review job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewJob"
...
components:
  schemas:
...
    PullRequest:
      type: object
      required: [id, number, title, author, repositoryId, repositoryName, filesChanged, commits, status, createdAt]
      properties:
        id:
          type: integer
        number:
          type: integer
        title:
          type: string
        author:
          type: string
        authorAvatar:
          type: ["string", "null"]
        repositoryId:
          type: integer
        repositoryName:
          type: string
        filesChanged:
          type: integer
        commits:
          type: integer
        status:
          type: string
          enum: [open, closed, merged]
        reviewStatus:
          type: ["string", "null"]
          enum: [pending, analyzing, completed, null]
        createdAt:
          type: string
    ReviewJob:
      type: object
      required: [jobId, pullRequestId, status]
      properties:
        jobId:
          type: string
        pullRequestId:
          type: integer
        status:
          type: string
        progress:
          type: integer
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentStatus"
        reviewId:
          type: integer
        error:
          type: string
    AgentStatus:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [waiting, running, completed, failed]
        progress:
          type: ["integer", "null"]
...