# lib/api-zod/src/generated/types/reviewJob.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Code Review Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface ReviewJob {
  jobId: string;
  pullRequestId: number;
  status: string;
  progress?: number;
  agents?: AgentStatus[];
  reviewId?: number;
  error?: string;
}